requirements.txt
BeautifulSoup
requests
mysql.connector